Conflicts:
	AlJeEngineVS2015/x64/Debug/AlJeEngineVS2015.ilk
	core/systems/Graphics/src/mesh.cpp

+++ core/systems/Graphics/src/CameraSystem.cpp
@@ -31,25 +31,31 @@ namespace AlJeEngine
     {
       for (auto &it : _entities)
       {
-        // do stuff with camera here
-        
+        //Getting the transform component on the object
         TransformPtr transform = it->GET_COMPONENT(Transform);
+        //Getting the camera component on the object
         CameraPtr camera = it->GET_COMPONENT(Camera);
 
         WindowSDLPtr window = GETSYS(WindowSDL);
-       
+        //Perspective of the camera, read more at 
+        //http://www.songho.ca/opengl/gl_projectionmatrix.html
         camera->_pespective = glm::perspectiveFov(camera->_fieldOfView, 
-                               static_cast<float>(transform->scale.x), 
-                               static_cast<float>(transform->scale.y), 
+                                                  transform->scale.x, 
+                                                  transform->scale.y, 
                                                   camera->_nearPlane, 
                                                   camera->_farPlane);
 
+        //Orthographic projection for the camera, read more at
+        //http://www.songho.ca/opengl/gl_projectionmatrix.html
         camera->_ortho = glm::ortho(transform->position.x - transform->scale.x * 0.5f,
                                     transform->position.x + transform->scale.x * 0.5f,
                                     transform->position.y - transform->scale.y * 0.5f,
                                     transform->position.y + transform->scale.y * 0.5f, 
                                     camera->_nearPlane,
                                     camera->_farPlane);
+
+        //View matrix for the camera, read more at 
+        //http://3dgep.com/understanding-the-view-matrix/
         camera->_viewMatrix = glm::lookAt(glm::vec3(transform->position.x, transform->position.y, -10.0f),
                                           camera->_target, 
                                           camera->_upVec);
